Workers/tests/submissions/Microsoft/postMessage_clone_port_error.htm

<!DOCTYPE html>
<html>
<head>
<title> postMessage(): cloning source port </title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
</head>
<body>
<div id=log></div>
<script>
    setup({timeout:500});
    var description = "Test Description: Throw an INVALID_STATE_ERR if the ports array in postMessage contains source port.";
   
    test(function()
    {
        var channel = new MessageChannel();
        channel.port1.start();
        
        assert_throws("INVALID_STATE_ERR", function()
        {
            channel.port1.postMessage("ports", [channel.port1]);
        });
    }, description);
</script>
</body>
</html>
